move quiz.js to new folder
